fix: defineSlot issues
fix: import interface from other vue file

chore: update packages

fix: no-dupe-keys issues
