+++ apps/docs/package.json
@@ -12,24 +12,24 @@
   },
   "devDependencies": {
     "@iconify-json/bi": "^1.1.17",
-    "@iconify-json/simple-icons": "^1.1.55",
-    "@rushstack/eslint-patch": "^1.3.0",
+    "@iconify-json/simple-icons": "^1.1.57",
+    "@rushstack/eslint-patch": "^1.3.2",
     "@tsconfig/node20": "^1.0.1",
-    "@types/node": "^20.2.5",
+    "@types/node": "^20.3.1",
     "@vue/eslint-config-prettier": "^7.1.0",
     "@vue/eslint-config-typescript": "^11.0.3",
     "@vue/tsconfig": "^0.4.0",
-    "@vueuse/core": "^10.1.2",
+    "@vueuse/core": "^10.2.0",
     "bootstrap": "^5.3.0",
     "bootstrap-vue-next": "workspace:^",
     "cross-env": "^7.0.3",
-    "eslint": "^8.42.0",
-    "eslint-define-config": "^1.20.0",
-    "eslint-plugin-vue": "^9.14.1",
+    "eslint": "^8.43.0",
+    "eslint-define-config": "^1.21.0",
+    "eslint-plugin-vue": "^9.15.0",
     "prettier": "^2.8.8",
     "typescript": "^5.1.3",
-    "unplugin-icons": "^0.16.2",
-    "vitepress": "1.0.0-beta.1",
+    "unplugin-icons": "^0.16.3",
+    "vitepress": "1.0.0-beta.3",
     "vue": "^3.3.4"
   },
   "lint-staged": {
