+++ apps/playground/package.json
@@ -15,14 +15,14 @@
   },
   "devDependencies": {
     "@tsconfig/node20": "^1.0.1",
-    "@types/node": "^20.2.5",
+    "@types/node": "^20.3.1",
     "@vitejs/plugin-vue": "^4.2.3",
     "@vue/tsconfig": "^0.4.0",
-    "bootstrap-styles": "^5.2.3",
-    "rollup-plugin-visualizer": "^5.9.0",
+    "bootstrap-styles": "^5.3.0",
+    "rollup-plugin-visualizer": "^5.9.2",
     "typescript": "^5.1.3",
     "vite": "^4.3.9",
-    "vue-tsc": "^1.6.5"
+    "vue-tsc": "^1.8.1"
   },
   "lint-staged": {
     "*.{js,ts,vue}": "eslint --cache --fix",
