+++ packages/bootstrap-vue-next/package.json
@@ -29,6 +29,7 @@
     "dev": "vite --port 5174",
     "build-only": "vite build",
     "type-check": "vue-tsc --noEmit -p tsconfig.app.json --composite false",
+    "type-check-emit": "vue-tsc --declaration --emitDeclarationOnly -p tsconfig.app.json --composite false",
     "preview": "vite preview",
     "build": "pnpm run type-check && pnpm run build-only",
     "lint": "eslint --ext .js,.ts,.vue --ignore-path ../../.gitignore --fix src",
@@ -46,42 +47,42 @@
   },
   "dependencies": {
     "@floating-ui/vue": "^1.0.1",
-    "@vueuse/core": "^10.1.2"
+    "@vueuse/core": "^10.2.0"
   },
   "devDependencies": {
-    "@floating-ui/core": "^1.2.6",
-    "@rushstack/eslint-patch": "^1.3.0",
+    "@floating-ui/core": "^1.3.1",
+    "@rushstack/eslint-patch": "^1.3.2",
     "@tsconfig/node20": "^1.0.1",
-    "@types/node": "^20.2.5",
+    "@types/node": "^20.3.1",
     "@vitejs/plugin-vue": "^4.2.3",
-    "@vitest/coverage-c8": "^0.31.4",
+    "@vitest/coverage-c8": "^0.32.2",
     "@vue/compiler-dom": "^3.3.4",
     "@vue/eslint-config-prettier": "^7.1.0",
     "@vue/eslint-config-typescript": "^11.0.3",
     "@vue/runtime-core": "^3.3.4",
     "@vue/shared": "^3.3.4",
-    "@vue/test-utils": "^2.3.2",
+    "@vue/test-utils": "^2.4.0",
     "@vue/tsconfig": "^0.4.0",
-    "@vueuse/shared": "^10.1.2",
+    "@vueuse/shared": "^10.2.0",
     "bootstrap": "^5.3.0",
-    "c8": "^7.14.0",
+    "c8": "^8.0.0",
     "csstype": "^3.1.2",
-    "eslint": "^8.42.0",
-    "eslint-define-config": "^1.20.0",
-    "eslint-plugin-vue": "^9.14.1",
+    "eslint": "^8.43.0",
+    "eslint-define-config": "^1.21.0",
+    "eslint-plugin-vue": "^9.15.0",
     "happy-dom": "^9.20.3",
     "jsdom": "^22.1.0",
     "prettier": "^2.8.8",
-    "rollup": "^3.23.0",
-    "rollup-plugin-visualizer": "^5.9.0",
-    "sass": "^1.62.1",
+    "rollup": "^3.25.1",
+    "rollup-plugin-visualizer": "^5.9.2",
+    "sass": "^1.63.6",
     "typescript": "^5.1.3",
     "vite": "^4.3.9",
     "vite-plugin-dts": "^2.3.0",
-    "vitest": "^0.31.4",
+    "vitest": "^0.32.2",
     "vue": "^3.3.4",
     "vue-router": "^4.2.2",
-    "vue-tsc": "^1.6.5"
+    "vue-tsc": "^1.8.1"
   },
   "repository": {
     "type": "git",
