+++ packages/bootstrap-vue-next/src/components/BLink/BLink.vue
@@ -28,37 +28,12 @@ import {useBooleanish} from '../../composables'
 import {collapseInjectionKey, navbarInjectionKey} from '../../utils'
 import {computed, getCurrentInstance, inject, type PropType, useAttrs} from 'vue'
 import type {RouteLocation, RouteLocationRaw} from 'vue-router'
-
-export interface BLinkProps {
-  active?: Booleanish
-  activeClass?: string
-  append?: Booleanish
-  disabled?: Booleanish
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  event?: string | any[]
-  href?: string
-  // noPrefetch: {type: [Boolean, String] as PropType<Booleanish>, default: false},
-  // prefetch: {type: [Boolean, String] as PropType<Booleanish>, default: null},
-  rel?: string
-  replace?: Booleanish
-  routerComponentName?: string
-  routerTag?: string
-  target?: LinkTarget
-  to?: RouteLocationRaw
-  variant?: ColorVariant | null
-  opacity?: 10 | 25 | 50 | 75 | 100 | '10' | '25' | '50' | '75' | '100'
-  opacityHover?: 10 | 25 | 50 | 75 | 100 | '10' | '25' | '50' | '75' | '100'
-  underlineVariant?: ColorVariant | null
-  underlineOffset?: 1 | 2 | 3 | '1' | '2' | '3'
-  underlineOffsetHover?: 1 | 2 | 3 | '1' | '2' | '3'
-  underlineOpacity?: 0 | 10 | 25 | 50 | 75 | 100 | '0' | '10' | '25' | '50' | '75' | '100'
-  underlineOpacityHover?: 0 | 10 | 25 | 50 | 75 | 100 | '0' | '10' | '25' | '50' | '75' | '100'
-  icon?: Booleanish
-}
+import type {BLinkProps} from '../../types/BLinkProps'
 
 // TODO this component will likely have an issue with inheritAttrs
 defineSlots<{
-  default?: Record<string, never>
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  default?: (props: Record<string, never>) => any
 }>()
 
 const props = withDefaults(defineProps<BLinkProps>(), {
