+++ packages/bootstrap-vue-next/src/components/BNavbar/BNavbarBrand.vue
@@ -7,10 +7,12 @@
 <script setup lang="ts">
 import {isLink, pluckProps} from '../../utils'
 import {computed} from 'vue'
-import BLink, {type BLinkProps} from '../BLink/BLink.vue'
+import BLink from '../BLink/BLink.vue'
+import type {BLinkProps} from '../../types/BLinkProps'
 
 defineSlots<{
-  default?: Record<string, never>
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  default?: (props: Record<string, never>) => any
 }>()
 
 const props = withDefaults(
