+++ packages/nuxt/package.json
@@ -33,20 +33,20 @@
     "bootstrap-vue-next": "workspace:^"
   },
   "dependencies": {
-    "@nuxt/kit": "^3.5.2"
+    "@nuxt/kit": "^3.6.0"
   },
   "devDependencies": {
     "@nuxt/eslint-config": "^0.1.1",
     "@nuxt/module-builder": "^0.4.0",
-    "@nuxt/schema": "^3.5.2",
-    "@rushstack/eslint-patch": "^1.3.0",
+    "@nuxt/schema": "^3.6.0",
+    "@rushstack/eslint-patch": "^1.3.2",
     "@vue/eslint-config-prettier": "^7.1.0",
     "bootstrap-vue-next": "workspace:^",
-    "eslint": "^8.42.0",
-    "eslint-define-config": "^1.20.0",
-    "nuxt": "^3.5.2",
+    "eslint": "^8.43.0",
+    "eslint-define-config": "^1.21.0",
+    "nuxt": "^3.6.0",
     "prettier": "^2.8.8",
-    "unimport": "^3.0.7",
+    "unimport": "^3.0.8",
     "vue": "^3.3.4"
   },
   "repository": {
